requirements.txt
PyExecJS
requests
loguru
python-dotenv
retry
openpyxl
Flask
Flask-CORS
gunicorn
openai
Pillow